+++ libraries/common/src/main/java/androidx/media3/common/MediaMetadata.java
@@ -84,6 +84,7 @@ public final class MediaMetadata {
     @Nullable private CharSequence author;
     @Nullable private CharSequence composer;
     @Nullable private CharSequence conductor;
+    @Nullable private CharSequence discSubtitle;
     @Nullable private Integer discNumber;
     @Nullable private Integer totalDiscCount;
     @Nullable private CharSequence genre;
@@ -129,6 +130,7 @@ public final class MediaMetadata {
       this.composer = mediaMetadata.composer;
       this.conductor = mediaMetadata.conductor;
       this.discNumber = mediaMetadata.discNumber;
+      this.discSubtitle = mediaMetadata.discSubtitle;
       this.totalDiscCount = mediaMetadata.totalDiscCount;
       this.genre = mediaMetadata.genre;
       this.compilation = mediaMetadata.compilation;
@@ -412,6 +414,14 @@ public final class MediaMetadata {
       return this;
     }
 
+    /** Sets the disc subtitle. */
+    @UnstableApi
+    @CanIgnoreReturnValue
+    public Builder setDiscSubtitle(@Nullable CharSequence discSubtitle) {
+      this.discSubtitle = discSubtitle;
+      return this;
+    }
+
     /** Sets the disc number. */
     @CanIgnoreReturnValue
     public Builder setDiscNumber(@Nullable Integer discNumber) {
@@ -606,6 +616,9 @@ public final class MediaMetadata {
       }
       if (mediaMetadata.conductor != null) {
         setConductor(mediaMetadata.conductor);
+      }
+      if (mediaMetadata.discSubtitle != null) {
+        setDiscSubtitle(mediaMetadata.discSubtitle);
       }
       if (mediaMetadata.discNumber != null) {
         setDiscNumber(mediaMetadata.discNumber);
@@ -1139,6 +1152,9 @@ public final class MediaMetadata {
   /** Optional conductor. */
   @Nullable public final CharSequence conductor;
 
+  /** Optional disc subtitle. */
+  @UnstableApi @Nullable public final CharSequence discSubtitle;
+
   /** Optional disc number. */
   @Nullable public final Integer discNumber;
 
@@ -1220,6 +1236,7 @@ public final class MediaMetadata {
     this.author = builder.author;
     this.composer = builder.composer;
     this.conductor = builder.conductor;
+    this.discSubtitle = builder.discSubtitle;
     this.discNumber = builder.discNumber;
     this.totalDiscCount = builder.totalDiscCount;
     this.genre = builder.genre;
@@ -1273,6 +1290,7 @@ public final class MediaMetadata {
         && TextUtils.equals(writer, that.writer)
         && TextUtils.equals(composer, that.composer)
         && TextUtils.equals(conductor, that.conductor)
+        && TextUtils.equals(discSubtitle, that.discSubtitle)
         && Objects.equals(discNumber, that.discNumber)
         && Objects.equals(totalDiscCount, that.totalDiscCount)
         && TextUtils.equals(genre, that.genre)
@@ -1314,6 +1332,7 @@ public final class MediaMetadata {
         writer,
         composer,
         conductor,
+        discSubtitle,
         discNumber,
         totalDiscCount,
         genre,
@@ -1359,6 +1378,7 @@ public final class MediaMetadata {
   private static final String FIELD_IS_BROWSABLE = Util.intToStringMaxRadix(32);
   private static final String FIELD_DURATION_MS = Util.intToStringMaxRadix(33);
   private static final String FIELD_SUPPORTED_COMMANDS = Util.intToStringMaxRadix(34);
+  private static final String FIELD_DISC_SUBTITLE = Util.intToStringMaxRadix(35);
   private static final String FIELD_EXTRAS = Util.intToStringMaxRadix(1000);
 
   // Use a fairly lenient threshold for sending byte array to legacy processes that don't support
@@ -1484,6 +1504,9 @@ public final class MediaMetadata {
     if (discNumber != null) {
       bundle.putInt(FIELD_DISC_NUMBER, discNumber);
     }
+    if (discSubtitle != null) {
+      bundle.putCharSequence(FIELD_DISC_SUBTITLE, discSubtitle);
+    }
     if (totalDiscCount != null) {
       bundle.putInt(FIELD_TOTAL_DISC_COUNT, totalDiscCount);
     }
@@ -1534,6 +1557,7 @@ public final class MediaMetadata {
         .setWriter(bundle.getCharSequence(FIELD_WRITER))
         .setComposer(bundle.getCharSequence(FIELD_COMPOSER))
         .setConductor(bundle.getCharSequence(FIELD_CONDUCTOR))
+        .setDiscSubtitle(bundle.getCharSequence(FIELD_DISC_SUBTITLE))
         .setGenre(bundle.getCharSequence(FIELD_GENRE))
         .setCompilation(bundle.getCharSequence(FIELD_COMPILATION))
         .setStation(bundle.getCharSequence(FIELD_STATION))