+++ libraries/extractor/src/main/java/androidx/media3/extractor/metadata/id3/TextInformationFrame.java
@@ -178,6 +178,9 @@ public final class TextInformationFrame extends Id3Frame {
         }
         // Don't set a numeric genre that we don't recognize.
         break;
+      case "TSST":
+        builder.setDiscSubtitle(values.get(0));
+        break;
       default:
         break;
     }