+++ libraries/extractor/src/main/java/androidx/media3/extractor/metadata/vorbis/VorbisComment.java
@@ -74,6 +74,9 @@ public final class VorbisComment implements Metadata.Entry {
           builder.setTotalTrackCount(totalTracks);
         }
         break;
+      case "DISCSUBTITLE":
+        builder.setDiscSubtitle(value);
+        break;
       case "DISCNUMBER":
         @Nullable Integer discNumber = Ints.tryParse(value);
         if (discNumber != null) {